Create prepare_multisect() method and test it.
For multisection, we will need a place in the object in which to keep track of
the results of tests as we run them one, algorithmically-select commit at a
time.  And as a baseline we need the results for the tests run at the first
and last commits in the commit range.  Create a 'bisected_outputs' element in
the object which, once this method runs, will be an array whose elements other
than the first and the last will be empty array references.