Start to refactor identify_transitions().
Now that we have a better idea of the various success conditions at each
stage, we'll rebuild the loop structures to test for success for each.

Start by identifying the overall success condition and storing data for it in
a hash.  Then refactor code for handling one target test file into new (mostly
internal) method multisect_one_target().  Start to test that message.

In the meantime, comment out code within identify_transitions() that has not
yet been restructured, so as to avoid spurious test failures.