Rename identify_transitions() to multisect_all_targets().
Rename prepare_multisect_hash() to prepare_for_multisection().

Call prepare_for_multisection() within multisect_all_targets().

Make it an error to call multisect_one_target() -- in the rare event that you
want to do that on a stand-alone basis -- without first having called
prepare_for_multisection().

This is needed because prepare_for_multisection() sets up the
'bisected_outputs' element in the object and populates it with the results of
the runs of tests at the first and last points in the commit range.