BuildTransitions: first attempt at "probe => 'stderr'"
Previously the only valid values for 'probe' were 'error' and 'warning'.
We used these when we want to identify transitions where there was a
change in either the errors or the warnings generated by a particular
C-compiler and version thereof.

However, there are also cases where a Perl-level warning is generated.
Since these warnings are not as rigidly formatted as those generated by
the C-compiler, it's not possible to write a regex to spot them.  So
what we have to do is capture everything written to STDERR while 'make'
is running and write that to a file.  We take a digest of that file and
use that digest to determine transitions which we then manually inspect.

TODO: Add test or example script.