Add program for testing changes in STDERR output from 'make'
At this stage of development programs that rely on BuildTransitions.pm
must pass a hashref to multisect_build.  But I need to explore whether
that should be passed in the hashref passed to new() instead.

Also, it has become apparent that the '-j${TEST_JOBS}' option for 'make'
should not be used in this library.  At least on  FreeBSD, when you use
that the output of 'make' changes dramatically.  Among other things,
file descriptor 2 fails to write anything to file; STDERR gets
intermixed with STDOUT.

TODO: POD for xt/stderr-transitions.pl.