Add t/012-build-transitions.t
This will test BuildTransitions.pm in a situation derived from
https://github.com/Perl/perl5/issues/16794.  With a particular, uncommon
set of configuration options, the build was broken (5 errors reported) at the point where
the multiconcat op was introduced.  A short time later 1 additional
build-time error was introduced.  (This program currently does not
extend to the point where the build was fixed.)