Revert "Remove xt/104-gcc-build-transitions-warnings.t"
This reverts commit ab8d09c3f98775e92c2728d0172d7da3c333f098.

This file was added to xt/ in 2019 after it (or one of its antecedents)
was used to debug https://github.com/Perl/perl5/issues/17347.  AFAICT
the debugging was done on Linux; the problem involved '-m32'
cross-compilation and a mix of 7 libraries.  Because, for various
reasons, the file was in xt rather than t/, it was never run as part of
'make test' on FreeBSD.  So the fact that it never worked on FreeBSD
meant that during the past week's development work I was never aware
that t/013 would not work on FreeBSD either.

Re-adding it to xt/ as a historical note.  I'll add some inline comments
that indicate why it's Linux-only.