Init: Guarantee tempfile is UNLINKed
We don't need it once we've seen a croak message.

t/002 thru t/004: Add tests for existence of directories.  Since we're
not doing any multisectioning in these files, we don't need to keep the
tempdirs assigned to 'outputdir' around.  So let's explicitly assign a
CLEANUP tempdir to 'outputdir' in creating the object.