binmode the tempfile, close the handle before hexdigest
We would like to delete as many tempfiles as possible created during
'make test'.  To do that we now use File::Temp's object-oriented
interface, with 'UNLINK => 1' provided to new().  But though that we
subsequently write the same content to the tempfile that we did with the
functional interface, hexdigest, for reasons not clear, calculates a
different digest.

Per discussion with Corion in https://perlmonks.org/?node_id=11136181,
we work around this in the test file.  After we create the File::Temp
object, we "binmode :raw" it, then close the handle after writing to it.

Also remove unused imports from two modules.